discord/requirements.txt
colorama==0.4.6
discord.py==2.1.0
python-dotenv==0.21.0
